DDF-04447 Move BlacklistItem model
