+++ ui/packages/catalog-ui-search/src/main/webapp/js/model/BlacklistItem.js
