+++ ui/packages/catalog-ui-search/src/main/webapp/js/model/User.js
@@ -26,7 +26,7 @@ const Alert = require('./Alert')
 const Common = require('../Common.js')
 const UploadBatch = require('./UploadBatch.js')
 const announcement = require('../../component/announcement/index.jsx')
-const BlackListItem = require('../../component/blacklist-item/blacklist-item.js')
+const BlackListItem = require('./BlacklistItem.js')
 const moment = require('moment-timezone')
 const Theme = require('./Theme.js')
 const ThemeUtils = require('../ThemeUtils.js')
